Adding pause state and played time

+++ UpdateUser.py
@@ -7,10 +7,11 @@ def update_user(user):
     try:
         conn = connect_to_db()
         cur = conn.cursor()
-        cur.execute("UPDATE users SET name = ?, email = ?, phone = ?, address = ?, country = ?, maxTimeAllowed = ?, WHERE user_id =?",
+        cur.execute("UPDATE users SET name = ?, email = ?, phone = ?, address = ?, country = ?, maxTimeAllowed = ?, "
+                    "playedTime = ?, WHERE user_id =?",
                      (user["name"], user["email"], user["phone"],
                      user["address"], user["country"],
-                     user["user_id"], user["maxTimeAllowed"]))
+                     user["user_id"], user["maxTimeAllowed"], user["isPaused"], user["playedTime"]))
         conn.commit()
         #return the user
         updated_user = get_user_by_id(user["user_id"])
@@ -28,8 +29,7 @@ def delete_user(user_id):
     message = {}
     try:
         conn = connect_to_db()
-        conn.execute("DELETE from users WHERE user_id = ?",
-                      (user_id,))
+        conn.execute("DELETE from users WHERE user_id = ?", (user_id,))
         conn.commit()
         message["status"] = "User deleted successfully"
     except:
